+++ src/Facades/ReferenceData/Sectors.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Digitonic\IexCloudSdk\Facades\ReferenceData;
+
+use Illuminate\Support\Facades\Facade;
+
+class Sectors extends Facade
+{
+    /**
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return \Digitonic\IexCloudSdk\ReferenceData\Sectors::class;
+    }
+}