+++ src/ReferenceData/Sectors.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Digitonic\IexCloudSdk\ReferenceData;
+
+use Digitonic\IexCloudSdk\Contracts\IEXCloud;
+use Digitonic\IexCloudSdk\Requests\BaseGet;
+
+class Sectors extends BaseGet
+{
+    const ENDPOINT = 'ref-data/sectors';
+
+    /**
+     * Create constructor.
+     *
+     * @param  IEXCloud  $api
+     */
+    public function __construct(IEXCloud $api)
+    {
+        parent::__construct($api);
+    }
+
+    /**
+     * @return string
+     */
+    protected function getFullEndpoint(): string
+    {
+        return self::ENDPOINT;
+    }
+}